+++ vacation-calculator/src/main/java/ru/vacation/calculator/vacation_calculator/services/VacationPayCalculationService.java
@@ -2,8 +2,10 @@ package ru.vacation.calculator.vacation_calculator.services;
 
 import ru.vacation.calculator.vacation_calculator.dto.VacationPayDto;
 
+import java.math.BigDecimal;
+
 public interface VacationPayCalculationService {
 
     // Расчет суммы отпускных
-    Double calculateVacationPay(VacationPayDto vacationPayDto);
+    BigDecimal calculateVacationPay(VacationPayDto vacationPayDto);
 }
